New Game Setup - Step 1
Created a blank game in Unreal Engine using a black map with C++ code supported. I also got a little ahead of myself before I created this repository and created a blank C++ Actor called Item. Right now this class is blank, but will be the parent class of all the items the Character will be able to pickup and interact with such as weapons, health items, etc.