Moving Objects With Code - Step 3
This part of the course was mostly education teaching how to move Actors across the scene using functions. For now our actor has a static mesh component as a pyramid and floats up and down using the Sin function. Next commit will focus on the Pawn class which will feature Pawns that will be permanent in the game.