The Pawn Class - Step 4
Created a Pawn class called bird and coded the class to accept movement inputs from the keyboard and mouse. Also created a KatanaGameMode to allow the Bird to the auto possessed at the start of the game. We will use this same game mode for development as the game goes. This section was practice for coding the main character in the next commit.