The Animation Blueprint - Step 5
Created an Anim Instance class (KatanaAnimInstance) to use as a base for the animation blueprint (ABP_Echo). So far we just have basic movement including idle, running and jumping. Also built a control rig so that the character feet look natural when standing on an incline. The knees bend so the feet can adjust to different angles of the surface the character is standing on. Not a lot of C++ code for this commit but Control Rigs do take a LOT of nodes in the blueprint!