Collision and Overlaps - Step 6
Added a sphere component to the Item class and coded two delegates in the class. One for triggering the beginning of overlapping and other for the ending of overlapping. I also downloaded a sword static mesh from Quixel Bridge and temporarily added that to the mesh component on the Item Blueprint. This is a good setup for the Weapon class which we will be coding in the next commit.