Attacking - Step 8
Created an Animation Montage to handle two different kinds of attacks when the left mouse button is clicked. Downloaded 2 animations from Mixamo to handle Arming and Disarming the Character's Weapon. Added an Action State for Attacking to prevent the Character from doing unnatural things (such has running) while attack animations are being played. Fixed an issue with the foot placement in the Idle/Equipped animation. Added sounds to both attack animations and also to the arming and disarming of the Weapon.